% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdapDictionaries-package.R
\docType{data}
\name{contractions}
\alias{contractions}
\title{Contraction Conversions}
\format{A data frame with 70 rows and 2 variables}
\usage{
data(contractions)
}
\description{
A dataset containing common contractions and their expanded form.
}
\details{
\itemize{ 
  \item contraction. The contraction word.
  \item expanded. The expanded form of the contraction.
}
}
\keyword{datasets}
