\docType{data}
\name{env.pol}
\alias{env.pol}
\title{Polarity Lookup Environment}
\format{A environment with  words and corresponding values.}
\usage{
data(env.pol)
}
\description{
A dataset containing a polarity lookup environment (see
\code{\link[qdap]{polarity}}).
}
\references{
Hu, M., & Liu, B. (2004). Mining opinion features in
customer reviews. National Conference on Artificial
Intelligence.

\url{http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\keyword{datasets}

