\docType{data}
\name{env.strength}
\alias{env.strength}
\title{Strength Lookup Environment}
\format{A environment with}
\usage{
data(env.strength)
}
\description{
A dataset containing a strength lookup environment.
}
\references{
\url{http://www.wjh.harvard.edu/~inquirer/inqdict.txt}
}
\keyword{datasets}

