\docType{data}
\name{env.syn}
\alias{env.syn}
\title{Synonym Lookup Environment}
\format{A environment with words and synonyms.}
\usage{
data(env.syn)
}
\description{
A dataset containing a synonym lookup environment (see
\code{SYNONYM}).
}
\references{
Scraped from:
\href{http://dictionary.reverso.net/english-synonyms/}{Reverso
Online Dictionary}. The word list fed to
\href{http://dictionary.reverso.net/english-synonyms/}{Reverso}
is the unique words from the combination of
\code{DICTIONARY} and \code{labMT}.
}
\keyword{datasets}

