% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{function.words}
\alias{function.words}
\title{Function Words}
\format{A vector with 321 elements}
\usage{
data(function.words)
}
\description{
A vector of function words from
\href{http://myweb.tiscali.co.uk/wordscape/museum/funcword.html}{John and Muriel Higgins's list}
used for the text game ECLIPSE.
}
\references{
\url{http://myweb.tiscali.co.uk/wordscape/museum/funcword.html}
}
\keyword{datasets}

