% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{interjections}
\alias{interjections}
\title{Interjections}
\format{A character vector with 139 elements}
\usage{
data(interjections)
}
\description{
A dataset containing a character vector of common interjections.
}
\references{
\url{http://www.vidarholen.net/contents/interjections/}
}
\keyword{datasets}

