% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{key.power}
\alias{key.power}
\title{Power Lookup Key}
\format{A hash key with power words.}
\usage{
data(key.power)
}
\description{
A dataset containing a power lookup key.
}
\references{
\url{http://www.wjh.harvard.edu/~inquirer/inqdict.txt}
}
\keyword{datasets}

