% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{key.syl}
\alias{key.syl}
\title{Syllable Lookup Key}
\format{A hash key with a modified DICTIONARY data set.}
\usage{
data(key.syl)
}
\description{
A dataset containing a syllable lookup key (see \code{DICTIONARY}).
}
\details{
For internal use.
}
\references{
\href{http://archive.ics.uci.edu/ml/machine-learning-databases/undocumented/connectionist-bench/nettalk/}{UCI Machine Learning Repository website}
}
\keyword{datasets}

