% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{key.syn}
\alias{key.syn}
\title{Synonym Lookup Key}
\format{A hash key with 10976 rows and 2 variables (words and synonyms).}
\usage{
data(key.syn)
}
\description{
A dataset containing a synonym lookup key.
}
\references{
Scraped from:
\href{http://dictionary.reverso.net/english-synonyms/}{Reverso Online Dictionary}.
The word list fed to \href{http://dictionary.reverso.net/english-synonyms/}{Reverso}
is the unique words from the combination of \code{DICTIONARY} and
\code{labMT}.
}
\keyword{datasets}

