% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Leveled_Dolch}
\alias{Leveled_Dolch}
\title{Leveled Dolch List of 220 Common Words}
\format{A data frame with 220 rows and 2 variables}
\usage{
data(Leveled_Dolch)
}
\description{
Edward William Dolch's list of 220 Most Commonly Used Words by reading level.
}
\details{
Dolch's Word List made up 50-75\% of all printed text in 1936.
\itemize{
  \item Word. The word
  \item Level. The reading level of the word
}
}
\references{
Dolch, E. W. (1936). A basic sight vocabulary. Elementary School
Journal, 36, 456-460.
}
\keyword{datasets}

