% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{negation.words}
\alias{negation.words}
\title{Negating Words}
\format{A vector with 23 elements}
\usage{
data(negation.words)
}
\description{
A dataset containing a vector of words that negate word meaning.
}
\details{
Valence shifters are words that alter or intensify the meaning of the polarized
words and include negators and amplifiers. Negators are, generally, adverbs
that negate sentence meaning; for example the word like in the sentence, "I do
like pie.", is given the opposite meaning in the sentence, "I do not like
pie.", now containing the negator not. Amplifiers are, generally, adverbs or
adjectives that intensify sentence meaning. Using our previous example, the
sentiment of the negator altered sentence, "I seriously do not like pie.", is
heightened with addition of the amplifier seriously.  Whereas de-amplifiers
decrease the intensity of a polarized word as in the sentence "I barely like
pie"; the word "barely" deamplifies the word like.
}
\keyword{datasets}

