% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qdapDictionaries-package.R
\docType{data}
\name{negative.words}
\alias{negative.words}
\title{Negative Words}
\format{A vector with 4776 elements}
\usage{
data(negative.words)
}
\description{
A dataset containing a vector of negative words.
}
\details{
A sentence containing more negative words would be deemed a negative sentence,
whereas a sentence containing more positive words would be considered positive.
}
\references{
Hu, M., & Liu, B. (2004). Mining opinion features in customer
reviews. National Conference on Artificial Intelligence.

\url{http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\keyword{datasets}

