% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qdapDictionaries-package.R
\docType{data}
\name{preposition}
\alias{preposition}
\title{Preposition Words}
\format{A vector with 162 elements}
\usage{
data(preposition)
}
\description{
A dataset containing a vector of common prepositions.
}
\keyword{datasets}

