% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Top25Words}
\alias{Top25Words}
\title{Fry's 25 Most Commonly Used English Words}
\format{A character vector with 25 elements}
\usage{
data(Top25Words)
}
\description{
A stopword list containing a character vector of stopwords.
}
\details{
Fry's Word List: The first 25 make up about one-third of all printed
material in English. The first 100 make up about one-half of all printed
material in English. The first 300 make up about 65\% of all printed
material in English."
}
\references{
Fry, E. B. (1997). Fry 1000 instant words. Lincolnwood, IL:
Contemporary Books.
}
\keyword{datasets}

