% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/view_data.R
\name{view_data}
\alias{view_data}
\title{List all data sets available in a qdapDictionaries}
\usage{
view_data()
}
\value{
Returns the data sets of qdapDictionaries as a dataframe.
}
\description{
Lists and describes all the data sets available in qdapDictionaries.
}
\examples{
view_data()
}
\seealso{
\code{\link[utils]{data}}
}

