% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qdapDictionaries-package.R
\docType{data}
\name{function.words}
\alias{function.words}
\title{Function Words}
\format{A vector with 350 elements}
\usage{
data(function.words)
}
\description{
A vector of function words from
\href{http://myweb.tiscali.co.uk/wordscape/museum/funcword.html}{John and Muriel Higgins's list}
used for the text game ECLIPSE.  The lest is augmented with additional
contractions from \code{\link[qdapDictionaries]{contractions}}.
}
\references{
\url{http://myweb.tiscali.co.uk/wordscape/museum/funcword.html}
}
\keyword{datasets}

