% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{GradyAugmented}
\alias{GradyAugmented}
\title{Augmented List of Grady Ward's English Words and Mark Kantrowitz's Names List}
\format{A vector with 120614 elements}
\usage{
data(GradyAugmented)
}
\description{
A dataset containing a vector of Grady Ward's English words augmented with
Mark Kantrowitz's names list, other proper nouns, and contractions.
}
\details{
A dataset containing a vector of Grady Ward's English words
augmented with proper nouns (U.S. States, Mark Kantrowitz's Names List, and
months) and contractions. That dataset is augmented for spell checking purposes.
}
\references{
Moby Thesaurus List by Grady Ward \url{http://www.gutenberg.org/etext/3202} \cr \cr
List of names from Mark Kantrowitz \url{http://www.cs.cmu.edu/afs/cs/project/ai-repository/ai/areas/nlp/corpora/names/}.
A copy of the \href{http://www.cs.cmu.edu/afs/cs/project/ai-repository/ai/areas/nlp/corpora/names/readme.txt}{README}
is available \href{http://www.cs.cmu.edu/afs/cs/project/ai-repository/ai/areas/nlp/corpora/names/readme.txt}{here}
per the author's request.
}
\keyword{datasets}

