% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{key.pol}
\alias{key.pol}
\title{Polarity Lookup Key}
\format{A hash key with words and corresponding values.}
\usage{
data(key.pol)
}
\description{
A dataset containing a polarity lookup key (see \code{\link[qdap]{polarity}}).
}
\references{
Hu, M., & Liu, B. (2004). Mining opinion features in customer
reviews. National Conference on Artificial Intelligence.

\url{http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\keyword{datasets}

