% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{key.strength}
\alias{key.strength}
\title{Strength Lookup Key}
\format{A hash key with strength words.}
\usage{
data(key.strength)
}
\description{
A dataset containing a strength lookup key.
}
\references{
\url{http://www.wjh.harvard.edu/~inquirer/inqdict.txt}
}
\keyword{datasets}

