% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdapDictionaries-package.R
\docType{package}
\name{qdapDictionaries}
\alias{qdapDictionaries}
\alias{package-qdapDictionaries}
\alias{qdapDictionaries-package}
\title{qdapDictionaries}
\description{
A collection of dictionaries and Word Lists to Accompany the qdap Package
}
