% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{qdapDictionaries}
\alias{package-qdapDictionaries}
\alias{qdapDictionaries}
\alias{qdapDictionaries-package}
\title{qdapDictionaries}
\description{
A collection of dictionaries and Word Lists to Accompany the qdap Package
}

