% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/view_data.R
\name{view_data}
\alias{view_data}
\title{List all data sets available in a \pkg{qdapDictionaries}}
\usage{
view_data(package = "qdapDictionaries")
}
\arguments{
\item{package}{The name of the package.}
}
\value{
Returns the data sets of \pkg{qdapDictionaries} as a dataframe.
}
\description{
Lists and describes all the data sets available in \pkg{qdapDictionaries}.
}
\examples{
view_data()
}
\seealso{
\code{\link[utils]{data}}
}

