\name{view_data}
\alias{view_data}
\title{List all data sets available in a qdapDictionaries}
\usage{
  view_data()
}
\value{
  Returns the data sets of qdapDictionaries as a dataframe.
}
\description{
  Lists and describes all the data sets available in
  qdapDictionaries.
}
\examples{
data_view()
}
\seealso{
  \code{\link[utils]{data}}
}

